requirements.txt
langchain
langchain-core
langchain-qdrant
langchain-google-genai
langchain-text-splitters

langgraph
pypdf
streamlit

python-dotenv
qdrant-client
sentence-transformers

# Testing dependencies
pytest
pytest-mock
pytest-asyncio
requests-mock